Agrego diferentes files de historia usuario
